Refactored and changed functionality to download based off country rankings
