--- cursor_state.txt
@@ -1,1 +1,0 @@
-Cursor(page=40)
