--- user_count_state.txt
@@ -1,1 +1,0 @@
-1950
